Adjust new output fields to facilitate radiation input variables.
This commit adjusts the set of `inRAD` variables that can be written
to file to reflect the current set of "Required" fields in the SCREAMv1
code.

It also adjusts where in the code the output is written ot hopefully better
represent the model state prior to any radiation tendency code being run.