Create a new rrtmgp_stand_alone test that initializes and runs through the AD.  This is in contrast to the existing rrtmgp test which initializes given a baseline of data and compares to baseline results from an rrtmgp testcase.
This commit changes the name of the old rrtmgp_stand_alone test to
rrtmgp_standalone_unit.

Introduces the new rrtmgp_standalone test which uses rrtmgp_init_ne2np4.nc
to initialize the run.

In keeping with the rest of the commits on this branch, the new stand alone
test is run up to the max MPI ranks.